requirements.txt
flask==1.1.2
Flask-API==2.0
Jinja2==2.11.2
Werkzeug==1.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
pymongo==3.11.0
pymysql==0.10.1
redis==3.5.3
python-automaton==1.3.1
konlpy==0.5.2
tweepy==3.10.0
nltk==3.5
scikit-learn==0.23.2
sklearn==0.0
transformers==4.18.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
torch==1.11.0
numpy==1.19.2
pandas==1.1.2